+++ tests/test_contrib_item.py
@@ -1,5 +1,4 @@
 import pickle
-from nose.tools import assert_equal
 from pomp.contrib.item import Item, Field
 
 
@@ -13,26 +12,26 @@ def test_item():
 
     # item as a dict
     i = TItem()
-    assert_equal(['f1', 'f2', 'f3'], list(i.keys()))
+    assert ['f1', 'f2', 'f3'] == list(i.keys())
 
     i.f1 = 1
     i.f3 = 3
-    assert_equal([1, None, 3], list(i.values()))
+    assert [1, None, 3] == list(i.values())
 
     # item as a object
     i = TItem('f1', f3='f3', f2='f2')
-    assert_equal(i.f1, 'f1')
-    assert_equal(i.f2, 'f2')
-    assert_equal(i.f3, 'f3')
-    assert_equal(['f1', 'f2', 'f3'], list(i.keys()))
+    assert i.f1 == 'f1'
+    assert i.f2 == 'f2'
+    assert i.f3 == 'f3'
+    assert ['f1', 'f2', 'f3'] == list(i.keys())
 
     # change attributes
     i.f1 = 'f1_new'
     i.f2 = 'f2_new'
-    assert_equal(i.f1, 'f1_new')
-    assert_equal(i.f2, 'f2_new')
-    assert_equal(i.f3, 'f3')
-    assert_equal(['f1', 'f2', 'f3'], list(i.keys()))
+    assert i.f1 == 'f1_new'
+    assert i.f2 == 'f2_new'
+    assert i.f3 == 'f3'
+    assert ['f1', 'f2', 'f3'] == list(i.keys())
 
     # pickling
-    assert_equal(i, pickle.loads(pickle.dumps(i)))
+    assert i == pickle.loads(pickle.dumps(i))